+++ product_code/authentication_service/api/src/test/scala/de/upb/cs/uc4/authentication/AuthenticationServiceStub.scala
@@ -3,28 +3,11 @@ package de.upb.cs.uc4.authentication
 import akka.{ Done, NotUsed }
 import com.lightbend.lagom.scaladsl.api.ServiceCall
 import de.upb.cs.uc4.authentication.api.AuthenticationService
-import de.upb.cs.uc4.authentication.model.AuthenticationRole.AuthenticationRole
-import de.upb.cs.uc4.authentication.model.{ AuthenticationRole, AuthenticationUser }
+import de.upb.cs.uc4.authentication.model.{ AuthenticationUser, JsonUsername }
 
 import scala.concurrent.Future
 
 class AuthenticationServiceStub extends AuthenticationService {
-
-  /** Checks if the username and password pair exists */
-  override def check(user: String, pw: String): ServiceCall[NotUsed, (String, AuthenticationRole)] = ServiceCall {
-    _ =>
-      if (user.contains("student")) {
-        Future.successful(user, AuthenticationRole.Student)
-      }
-      else {
-        if (user.contains("lecturer")) {
-          Future.successful(user, AuthenticationRole.Lecturer)
-        }
-        else {
-          Future.successful("admin", AuthenticationRole.Admin)
-        }
-      }
-  }
 
   /** Sets the authentication data of a user */
   override def allowVersionNumber: ServiceCall[NotUsed, Done] = ServiceCall { _ => Future.successful(Done) }
@@ -37,4 +20,16 @@ class AuthenticationServiceStub extends AuthenticationService {
 
   /** This Methods needs to allow a GET-Method */
   override def allowedPut: ServiceCall[NotUsed, Done] = ServiceCall { _ => Future.successful(Done) }
+
+  /** Logs a user in and returns a refresh and a login token in the header */
+  override def login: ServiceCall[NotUsed, Done] = ServiceCall { _ => Future.successful(Done) }
+
+  /** Allows GET */
+  override def allowedGet: ServiceCall[NotUsed, Done] = ServiceCall { _ => Future.successful(Done) }
+
+  /** Generates a new login token from a refresh token */
+  override def refresh: ServiceCall[NotUsed, JsonUsername] = ServiceCall { _ => Future.successful(JsonUsername("MOCK")) }
+
+  /** Logs the user out */
+  override def logout: ServiceCall[NotUsed, Done] = ServiceCall { _ => Future.successful(Done) }
 }