+++ product_code/deploy/deploy.sh
@@ -52,17 +52,15 @@ echo
 echo "##############################"
 echo "#     Starting Services      #"
 echo "##############################"
-kubectl create secret generic user-application-secret --from-literal=secret="test" -n uc4-lagom
+kubectl create secret generic application-secret --from-literal=secret="test" -n uc4-lagom
+
 kubectl create secret generic postgres-user --from-literal=username="uc4user" --from-literal=password="uc4user" -n uc4-lagom
 kubectl apply -f services/user.yaml
 
-kubectl create secret generic authentication-application-secret --from-literal=secret="test" -n uc4-lagom
 kubectl create secret generic postgres-authentication --from-literal=username="uc4authentication" --from-literal=password="uc4authentication" -n uc4-lagom
 kubectl apply -f services/authentication.yaml
 
-kubectl create secret generic course-application-secret --from-literal=secret="test" -n uc4-lagom
 kubectl create secret generic postgres-course --from-literal=username="uc4course" --from-literal=password="uc4course" -n uc4-lagom
 kubectl apply -f services/course.yaml
 
-kubectl create secret generic matriculation-application-secret --from-literal=secret="test" -n uc4-lagom
 kubectl apply -f services/matriculation.yaml