+++ product_code/authentication_service/CHANGELOG.md
@@ -1,6 +1,10 @@
 # [v.0.6.1](https://github.com/upb-uc4/University-Credits-4.0/compare/authentication-v0.6.0...authentication-v0.6.1) (2020-XX-XX)
 ## Feature
- - Added custom deserialization exceptions 
+ - Added custom deserialization exceptions
+ - Switch to JWT Token
+    - Added login endpoint which uses basic to create a refresh and a login token
+    - Added refresh endpoint to create a new login token with the refresh token
+ - The authentication service is now only needed for the first login 
 ## Refactor
  - Added ServiceStub for testing to reduce code duplication
  - Added default AuthenticationUsers for easier maintainability in testing